+++ packages/components/CHANGELOG.md
@@ -14,6 +14,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0), 
 
 ### [Unreleased]
 
+### [3.28.0-rc.0] - 2025-04-11
+
 #### Added
 
 - `Input Password`: ([#3763](https://github.com/porsche-design-system/porsche-design-system/pull/3763))